server/requirements.txt
qiskit-aer-gpu
more-itertools
grpcio
requests
qiskit-ibm-runtime